feat: Add lazy loading for images and update gitignore
The code changes include adding the `.idea/` directory to the `.gitignore` file and updating the `views.py` file in the `cart` app. Additionally, the `ProductConfig` class in the `product/apps.py` file was renamed from `ProductsConfig`. The `urls.py` file in the `product` app was also modified to include new paths for `ProductImageView`. Lastly, the `models.py` file in the `cart` app was updated to fix a typo and add a new model for `CartItem`.

feat: add product-image uploads and serialization